refactor(ors_control_app): extract routes/matrix
Move all 13 matrix endpoints + their internal helpers into routes/matrix.ts
as a single Express Router. Mounted via app.use(createMatrixRouter()).

Endpoints moved:
- /api/matrix/regions, /api/matrix/road-filter-available
- /api/matrix/cost-estimate, /api/matrix/existing
- /api/matrix/build, /api/matrix/status, /api/matrix/cancel
- /api/matrix/inventory, DELETE /api/matrix/:region/:profile/:resolution
- /api/matrix/:region/:profile/:resolution/restore (POST)
- /api/matrix/viewer-inventory, /api/matrix/random-origin
- /api/matrix/all-hexes, /api/matrix/reachability, /api/matrix/ring-stats

Internal helpers parseViewerTableName, getViewerInventory,
validateViewerTable moved as nested functions inside the router factory.

Lines down 1,650 -> 884 in server/index.ts (a 4.2x reduction from the
original 3,726-line monolith).

Phase 1 of holistic-control-app-refactor-final, route extraction step 7.
No behaviour change. tsc passes cleanly.
